Backend_work/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
cloudinary==1.44.0
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
dj-database-url==3.0.0
Django==5.2.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
ngrok==1.4.0
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-decouple==3.8
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
zope.interface==7.2
google-generativeai==0.3.0
google-api-python-client==2.130.0
